remove lspconfig and use vim 0.8 apis directly
